+++ terraform/main.go
@@ -2,7 +2,8 @@ package main
 
 import (
 	"flexbot/terraform/provider"
-	"github.com/hashicorp/terraform/plugin"
+	//"github.com/hashicorp/terraform/plugin"
+	"github.com/hashicorp/terraform-plugin-sdk/plugin"
 )
 
 func main() {