+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.4.2 (November 20, 2020)
+
+FEATURES:
+
+* This release is all about improved Terraform provider v1.4.2
+* See Terraform CHANGELOG for more details.
+
+ENHANCEMENTS:
+
+BUG FIXES:
+
 ## 1.4.1 (November 20, 2020)
 
 FEATURES: